requirements.txt
websocket-client>=1.6.0
requests>=2.31.0
py-clob-client>=0.18.0
python-dotenv>=1.0.0
websockets>=11.0.0
aiohttp>=3.9.0
